+++ src/components/generatorTourDetail.js
